toteutettu yksinkertainen admin kirjautuminen

+++ client/src/JarjestajienMuokkaus.js
@@ -1,27 +1,55 @@
-import React, { useRef } from 'react'
+import React, { useRef, useContext } from 'react'
+import Context from './Context'
+import jwtIsValid from './helpers/jwtIsValid'
 import axios from 'axios'
 
 function JärjestäjienMuokkaus({ kilpailu, setKilpailu, aktiivinenKausi }) {
+  const { setKirjauduttu } = useContext(Context)
   const järjestäjänNimiInput = useRef(null)
 
   function lisääJärjestäjä() {
     const nimi = järjestäjänNimiInput.current.value
     järjestäjänNimiInput.current.value = ''
+
+    if (!jwtIsValid()) {
+      localStorage.removeItem('jwt')
+      axios.defaults.headers['Authorization'] = null
+      setKirjauduttu(false)
+      return
+    }
+
     axios.post(`api/jarjestajat/${aktiivinenKausi.id}/${kilpailu._id}`, { nimi: nimi })
       .then(vastaus => {
         setKilpailu(vastaus.data)
       })
       .catch(err => {
+        if (err.response.status === 401) {
+          setKirjauduttu(false)
+          localStorage.removeItem('jwt')
+          axios.defaults.headers.common['Authorization'] = null
+        }
         console.log(err)
       })
   }
 
   function poistaJärjestäjä(järjestäjäId) {
+    if (!jwtIsValid()) {
+      localStorage.removeItem('jwt')
+      axios.defaults.headers['Authorization'] = null
+      setKirjauduttu(false)
+      return
+    }
+
     axios.delete(`api/jarjestajat/${aktiivinenKausi.id}/${kilpailu._id}/${järjestäjäId}`)
       .then(vastaus => {
         setKilpailu(vastaus.data)
       })
       .catch(err => {
+        if (err.response.status === 401) {
+          setKirjauduttu(false)
+          localStorage.removeItem('jwt')
+          axios.defaults.headers.common['Authorization'] = null
+        }
         console.log(err)
       })
   }
