+++ client/src/KilpailunMuokkaus.js
@@ -1,16 +1,30 @@
-import React, { useState, useRef } from 'react'
+import React, { useState, useEffect, useRef, useContext } from 'react'
 import axios from 'axios'
 import moment from 'moment'
+import Context from './Context'
+import jwtIsValid from './helpers/jwtIsValid'
 import SarjojenMuokkaus from './SarjojenMuokkaus'
 import JärjestäjienMuokkaus from './JarjestajienMuokkaus'
 import ManuaalistenPisteidenMuokkaus from './ManuaalistenPisteidenMuokkaus'
 
 function KilpailunMuokkaus({ aktiivinenKausi, kilpailu, setKilpailu, päivitäKausienJaKilpailujenNimet,
-  kilpailunMuokkaus, setKilpailunMuokkaus }) {
+  setKilpailunMuokkaus }) {
+  const { setKirjauduttu } = useContext(Context)
+
   const [muokattava, setMuokattava] = useState(null)
 
   const kilpailunNimiInput = useRef(null)
   const pvmInput = useRef(null)
+
+  useEffect(() => {
+    return () => setKilpailunMuokkaus(false)
+    // eslint-disable-next-line
+  }, [])
+
+  useEffect(() => {
+    setKirjauduttu(jwtIsValid())
+    // eslint-disable-next-line
+  }, [muokattava])
 
   function tyhjennäTekstit() {
     kilpailunNimiInput.current.value = ''
@@ -21,30 +35,56 @@ function KilpailunMuokkaus({ aktiivinenKausi, kilpailu, setKilpailu, päivitäKa
     let pyyntöBody = {}
     if (kilpailunNimiInput.current.value) pyyntöBody.nimi = kilpailunNimiInput.current.value
     if (pvmInput.current.value) pyyntöBody.pvm = moment(pvmInput.current.value, 'DD.MM.YYYY')
+    kilpailunNimiInput.current.value = ''
+    pvmInput.current.value = ''
+
+    if (!jwtIsValid()) {
+      localStorage.removeItem('jwt')
+      axios.defaults.headers['Authorization'] = null
+      setKirjauduttu(false)
+      return
+    }
 
     axios.put(`api/kilpailut/${aktiivinenKausi.id}/${kilpailu._id}`, pyyntöBody)
       .then(vastaus => {
         setKilpailu(vastaus.data)
       })
       .catch(err => {
+        if (err.response.status === 401) {
+          setKirjauduttu(false)
+          localStorage.removeItem('jwt')
+          axios.defaults.headers.common['Authorization'] = null
+        }
         console.log(err)
       })
-
-    setKilpailunMuokkaus(false)
   }
 
   function poistaKilpailu() {
+    if (!jwtIsValid()) {
+      localStorage.removeItem('jwt')
+      axios.defaults.headers['Authorization'] = null
+      setKirjauduttu(false)
+      return
+    }
+
     setKilpailunMuokkaus(false)
     axios.delete(`api/kilpailut/${aktiivinenKausi.id}/${kilpailu._id}`)
       .then(vastaus => {
         päivitäKausienJaKilpailujenNimet()
       })
       .catch(err => {
+        if (err.response.status === 401) {
+          setKirjauduttu(false)
+          localStorage.removeItem('jwt')
+          axios.defaults.headers.common['Authorization'] = null
+        }
         console.log(err)
       })
   }
 
   function asetaMuokattava(asetettava) {
+    setKirjauduttu(jwtIsValid())
+    
     if (muokattava === asetettava) {
       setMuokattava(null)
     } else {
